Make ldirectord's version valid
Previously the version was set to the (doctored) CVS revision
using SCCS keyword substitution. This is a bit tricker to achive
using Mercurial, where the code is now maintained. So just us a simpler
scheme of 1.186-ha-VERSION. Where 1.186 is the most recent version
that was released from CVS. And VERSION is the linux-ha version,
currently 2.0.9.

Signed-off-by: Simon Horman <horms@verge.net.au>