[DEBIAN] Rename the heartbeat-2 source package heartbeat
I was going to wait until the next linux-ha release before doing this, but
Debian is currently in a state of mass change in the aftermath of releasing
etch. So it seems like a good time to make this change.

A description of the motivation for this change is in debian/changelog

Signed-off-by: Simon Horman <horms@verge.net.au>