crmd: Novell Bug #286393 - Create resources, if necessary, in the lrmd so they can be stopped.
If a resource is "cleaned up" while the it is being being acted on, then it is
possible that we will send a STOP action at a point when it no longer exists in
then lrmd.
This patch removes the assumption that this will never happen and allows sane
handling of the situation.