Medium: crmd: Prevent shutdowns initiated immediately after a node is removed from the cluster with hb_delnode from stalling.
In such cases, we get a OC_EV_MS_PRIMARY_RESTORED event from the ccm and
although we update current_ccm_membership_id, do_pe_invoke_callback()
checks the actual cache (which has the same membership but the old id).

The easiest way to be correct is to just update the cache too.