ldirectord: SimpleTCP service check documentation We're only in development mode here right now, but it's stood up well under the testing we have done including some load testing. The code is pretty much just the check_sip code, with the request swapped out.
Patch attached with some tiny bits of documentation. Sorry about the
previous patch. hopefully this one will come through correctly.

One area I see potential improvement is the \n handling - I can see the
need for at least \r as well, probably others. I was looking for a quick
way to have perl parse the line as it does string literals in code so that
all the same perl escape syntax could be used (enabling embedded nulls or
other binary data), but didn't see an easy way and I didn't need that for
the current project so I left it. If someone knows a simple, safe and
secure way to enable that, I think it would really improve this to where it
can be used in as many circumstances as possible.

Signed-off-by: Aaron J. Linnen <aaron@lfchosting.com>
Signed-off-by: Simon Horman <horms@verge.net.au>