Remove bashisms in ocf-tester, hb_report and heartbeat's init script
Debian bug #464985
See: http://bugs.debian.org/#464985

In the case of the use of "echo -e" in ocf-tester, the -e flag doesn't
exist in dash and it interpates escaape sequences by default.  Bash on the
other hand has the -e flag and may or may not interprate escape sequences
by default. In any case, it seems that it is never called with a string
that includes escape sequences, so it is safe just to remove -e.

In the case of the use of cnt=$((cnt + 1) in hb_report,
it seems that cnt=$(($cnt + 1) is the non-bashish way to do this.

And in the case of the use of cmd &> /dev/null in heartbeat's init
script, it seems that cmd >/dev/null 2>&1 is the non-bashish way
to do this.

Signed-off-by: Simon Horman <horms@verge.net.au>