build: Don't distribute cim_provider As configure.in currently stands, as of the hg dev tree 54723736ab18, make dist fails because autogenerated files in cim/, in parcicular cim/Makefile.am, are not created.
Andrew Beekhof tells me that this is because cim_provider can't build
without the crm, which is now in the pacemaker tree along with
cim_provider.

This fairly minimal patch allows make dist to succeed by not
distributing cim/. I have found that building the resulting
tarball also succeeds.

A fuller patch would either fix cim_provider so that it can compile,
or remove it completely from cim_provider/, configure.in and
heartbeat.rpm.spec. I don't feel that it is within my jurisdiction
to make either change at this time.

Signed-off-by: Simon Horman <horms@verge.net.au>