build: Don't distribute tsa_plugin As configure.in currently stands, as of the hg dev tree 54723736ab18, make dist fails because autogenerated files in tsa_plugin/, in parcicular tsa_plugin/Makefile.am, are not created.
Andrew Beekhof tells me that this is because tsa_plugin can't build
without the crm, which is now in the pacemaker tree along with
tsa_plugin.

This fairly minimal patch allows make dist to succeed by not
distributing tsa_plugin/. I have found that building the resulting
tarball also succeeds.

A fuller patch would either fix tsa_plugin so that it can compile,
or remove it completely from tsa_plugin/, configure.in and
heartbeat.rpm.spec. I don't feel that it is within my jurisdiction
to make either change at this time.

Signed-off-by: Simon Horman <horms@verge.net.au>