Medium: safely create tempoary files After my previous post, I realised that there are lots of other places in the code where mktemp is used and more importantly, alternate unsafe code is used. So I have had a crack at cleaning things up.
------------------------------------------------------------------------

There are currently at least two problems with maketempfile.
Firstly, there is a race in the following constrct:

   rm -f "$F"; touch "$F"

As an attacker could potitinally create a symlink to "$F" between
the call to rm and the call to touch.

Secondly the use of $RANDOM appears to be a bashism.
On dash its usage in BasicSanityCheck appears to evaluate to
the empty string. See Debian Bug #489607, http://bugs.debian.org/489607

This patch takes the approach of using mkdir, which is atomic,
to create a safe place to store the logfile the tempoary directory
that is used by the filesystem check.

The patch also makes sure that the return value is checked
and the script exits cleanly if an error occurs.

This code is always used, there seems no point is providing
a robust fallback to mktemp - if its robust, it can be used :-)

For a discussion of creating tempoary files in shell see
http://www.linuxsecurity.com/content/view/115462/81/

Signed-off-by: Simon Horman <horms@verge.net.au>