medium: debian: make build-esential an alternate build-dependancy of libopenipmi-dev
This is a bit of a hack to make the openbuild system happy.
The problem is that Ubuntu 8.04 doesn't have libopenipmi-dev other
than in universe. But Debian does.

With this change the same control file should be usable to by
the build system for both Debian and Ubuntu.

Libnet is another similar problem, but it will be dealt with
separately.

This hack should probably be trimmed out of any official packages
that are uploaded to debian.

Signed-off-by: Simon Horman <horms@verge.net.au>