ldirectord: add option for executing command on fallback fallbackcommand = "path to script"
If this directive is defined, the supplied script is executed whenever
all real servers for a virtual server are down or when the first real
server comes up again. In the first case, it is called with "start" as
its first argument, in the latter with "stop".

===================

The current ldirectord does a fallback on startup and then only stops
the fallback once the first real server is checked and up. Because this
new option is tied to the existing code, the fallback script is also
started and stopped once at the startup of ldirectord, which might be
annoying for some users. That's why I'm not sure if there is general
interest in including it like this. I didn't dig deeper into the
ldirectord code to prevent this startup behavior from happening because
we actually want this behavior.

Signed-off-by: Simon Horman <horms@verge.net.au>