ldirectord: reap child processes The signal handling has been set up to auto-reap children, which is needed for the case where they time out and thus aren't reaped by the waitpid() that is called inside system().
However due to the wonders of perl, setting autoreap actually
changes the return value of waitpid() from > 0 to -1. This
breaks system() and is the root cause of the problem you are seeing.

My proposed fix is below. It sets up a signal handler that
just reaps the childern as neccessary. And it does this globally,
as there seems to be no good reason not to.

An alternative, which is just a work-around, is to simply remove
the "local $SIG{CHLD} = 'IGNORE';" which appears around line 2290.
This, however, will lead to zombies of your check process ever
times out.

Cc: Julius Volz <juliusv@google.com>
Cc: Stefan Ott <ott@id.unibe.ch>
Signed-off-by: Simon Horman <horms@verge.net.au>