Medium: RA: iSCSILogicalUnit: add support for SCSI ID, SCSI SN, Vendor ID, and Product ID
It seems nice to be able to set these in an implementation-independent
way. These VPD attributes could already be set previously by using the
"params" instance attribute, but that approach required observing
that, for example, the SCSI ID is named "ScsiId" in IET while it's
"scsi_id" in tgt. Now, you can just set "scsi_id" and the RA will
translate to the appropriate target parameter, as per the
implementation.

Being able to set the SCSI ID is specifically helpful as some iSCSI
initiator implementations rely on consistent SCSI IDs for smooth
target failover.