Low: RA: iSCSILogicalUnit: set default for SCSI SN, truncate SCSI ID default to 24 bytes
SCSI IDs are limited to 24 bytes, thus if $OCF_RESOURCE_INSTANCE is
longer than that, truncate it for $OCF_RESKEY_scsi_id_default.

Also, set a cluster-wide unique, failover persistent default for the
SCSI serial number. I choose the first 8 bytes of an MD5 hash of
$OCF_RESOURCE_INSTANCE. SCSI SNs can actually be up to 16 bytes long,
but 8 bytes should be sufficient to provide uniqueness.

All of this, of course, is to ensure smooth failover for iSCSI
initiators that read the SCSI ID and SN upon reconnect after a
connection interruption.