Medium: init script: don't touch /var/run/ppp.d
This line has been in the init script since the beginning of time
(it's actually in the very first version controlled incarnation in
the hg history), and apparently no-one ever bothered to remove it.
Whatever the original motivation, it's definitely not our business
to touch upon a directory owned by another package.