High: mark node dead on very late heartbeat
In some virtualized or other "special" environments,
sometimes heartbeat (or the whole VM) won't be scheduled
for more than deadtime.

It would then only detect a "Late heartbeat",
but no node-dead event.

If some of the other nodes had seen this node as dead,
this could confuse the membership layer (ccm):
some nodes would have an updated membership,
the "not scheduled" ones would retain the "old" membership.

Detect those "*very* late heartbeats", and mark those nodes
as dead-then-returning.

If we detect *ourselves* as dead,
try to get away with just tickling the ccm.