Medium: fix usage of SO_REUSEPORT in ucast sockets
Linux learned SO_REUSEPORT only with kernel 3.9,
but some linux headers already define SO_REUSEPORT.
Which, on older kernels, will result in ENOPROTOOPT,
"Protocol not available".

Failure to set SO_REUSEPORT is NOT critical in general.
It *may* be a problem on certain BSDs,
with more than two nodes, all using ucast.

Refusing to start because of failure to set SO_REUSEPORT is
not helpful for the vast majority of the clusters out there.

While at it, downgrade "critical" log messages to warnings
in non-fatal situations.