fix ambiguous "GPL v 2.1"
Apparently, back in 2000, at least heartbeat/api_test.c
was first part of a "batch tag" with the LGPLv2.1 stanza,
then was manually "re-tagged" as GPLv2 (because, well,
not a library, but a test program).

But the version (2.1) was not corrected to 2 at the same time.
There is no GPL v 2.1, though.

From there this "GPLv2.1" propagated to ipfail.c,
and later attrd.c, and other source code files.

See also:
  Date: Fri, 18 Mar 2016 16:16:01 +0100
  From: Lars Ellenberg <lars.ellenberg@linbit.com>
  To: developers@clusterlabs.org
  Subject: Re: [ClusterLabs Developers] [booth][sbd] GPLv2.1+ clarification request
  Message-ID: <20160318151601.GE4437@soda.linbit>
  http://clusterlabs.org/pipermail/developers/2016-March/000180.html
  http://thread.gmane.org/gmane.comp.clustering.clusterlabs.devel/133/focus=141
  https://www.mail-archive.com/developers@clusterlabs.org/msg00137.html


Affected:
  "GPLv2.1" -> LGPL v 2.1
	lib/plugins/HBcomm/bcast.c

  "GPLv2.1" -> GPL v2
	contrib/drbd-outdate-peer/dopd.c
	contrib/drbd-outdate-peer/dopd.h
	contrib/drbd-outdate-peer/drbd-peer-outdater.c
	contrib/ipfail/ipfail.c
	contrib/ipfail/ipfail.h
	lib/hbclient/api_test.c
	lib/lrm/racommon.c